api: refactor code
Changes:

- move units methods to methods package
- move cache initialization to main to avoid import loop
